requirements.txt
python-dotenv==1.0.0
pandas==2.1.1
sqlalchemy==2.0.21
psycopg2-binary==2.9.9
openpyxl==3.1.2
boto3==1.28.36
python-dateutil==2.8.2
logging==0.5.1.2
